[#7] refator : 상세정보페이지를 2개의 components로 분리
기존 상세정보페이지인 Detail에서
components/DetailInfo
components/DetailMain
으로 두개의 컴포넌트로 분리했습니다.
